editorial: Add xref to respecConfig, remove duplicate terms from Terminology.
Rather than defining multiple terms that come from other specs here, use the
`xref` configuration option to automatically reference most specs (and
explicitly reference those that are not added automatically).

This allows us to remove most terms from the Terminology section (including
some which we were not even referencing anymore) except for "fires an
event". DOM only defines it as "fire an event", and it is better to adjust
the spec text in another commit rather than fix it here.

Some definitions have changed as a consequence:
* "Promise": instead of referencing the ECMAScript definition of a Promise
  object, we now reference the Web IDL one.
* "policy-controlled feature" and "default allowlist": we are not
  referencing the TR version of the Permissions Policy spec rather than the
  latest ED.