editorial: Modernize the section covering Navigator.getBattery().
Replace most of the prose with internal slots and reference more terms in
the `getBattery()` algorithm.

The idea was to adopt more modern spec writing practices and make the text
stricter, but make no user-visible changes. This includes preserving some
oddities in the spec, such as always returning the same promise in
`getBattery()`.